+++ config/router.go
@@ -1,9 +1,7 @@
 package config
 
 import (
-	bookHandler "github.com/Aorjoa/bookstore/module/book/handler"
-	bookRepository "github.com/Aorjoa/bookstore/module/book/repository"
-	bookUseCase "github.com/Aorjoa/bookstore/module/book/usecase"
+	"github.com/Aorjoa/bookstore/module/book"
 	"github.com/gofiber/fiber/v2"
 	"gorm.io/gorm"
 )
@@ -15,8 +13,8 @@ func Router(db *gorm.DB) *fiber.App {
 	v1 := app.Group("/api/v1")
 
 	bg := v1.Group("/books")
-	br := bookRepository.NewBookRepository(db)
-	buc := bookUseCase.NewBookUseCase(br)
-	bookHandler.NewBookHttpHandler(bg, buc)
+	br := book.NewBookRepository(db)
+	buc := book.NewBookUseCase(br)
+	book.NewBookHttpHandler(bg, buc)
 	return app
 }
