+++ module/book/create_book_request.go
@@ -1,4 +1,4 @@
-package dto
+package book
 
 type CreateBookRequest struct {
 	Name     *string `json:"name"`
