--- module/book/entity/book.go
@@ -1,13 +1,0 @@
-package entity
-
-import (
-	"gorm.io/gorm"
-)
-
-type Book struct {
-	gorm.Model
-	Name     *string
-	ISBN     *string `gorm:"unique"`
-	Language *string
-	Status   *bool
-}
